update to use auth me endpoint
