+++ Client/Shared/MainLayout.razor
@@ -5,12 +5,12 @@
 </div>
 
 <div class="main">
-    <div class="top-row px-4">
+    @* <div class="top-row px-4">
         <a href="/login" target="_blank" class="ml-md-auto">Login</a>
     </div>
     <div class="top-row px-4">
         <a href="/logout" target="_blank" class="ml-md-auto">Logout</a>
-    </div>
+    </div> *@
 
     <div class="content px-4">
         @Body
