--- Shared/User/UserInfo.cs
@@ -1,7 +1,0 @@
-namespace BlazorApp.Shared.User
-{
-    public class UserInfo
-    {
-        public string Name { get; set; }
-    }
-}
